Feature/Implement DAG logic with Topological sorting
Basic Stack and Set data structures are used to form the components of the graph. The graph itself is implemented in a struct maintaining an adjecency list. The Tarjan algorithm is implemented to find the strongly connected components in a graph. This is then used to detect cycles. Basic departure time based topological sorting is used to determine the absolute order in the graph.